[#10760] Never use setuptools_scm except when importing astropy from within a checkout of the git repository
This adds a special package astropy._dev which is never installed, and only
exists when developing Astropy from within the source repo.  It is ommitted
from distributions such as tarballs and wheels.

astropy/version.py becomes a file checked into the repo, and the file
simplified.

astropy/version.py will use the version from astropy._dev._scm_version if
it exists (generated at runtime by setuptools_scm).  Otherwise it will use
the static version from astropy/_version.py.
