+++ .gitignore
@@ -15,10 +15,10 @@ __pycache__
 
 # Other generated files
 MANIFEST
-astropy/version.py
 astropy/cython_version.py
 astropy/wcs/include/wcsconfig.h
 astropy/_erfa/core.py
+astropy/_version.py
 
 # Sphinx
 _build
