+++ MANIFEST.in
@@ -25,6 +25,10 @@ recursive-include cextern *
 recursive-include scripts *
 recursive-include static *
 
+# This subpckage is only used in development checkouts and should not be
+# included in built tarballs
+prune astropy/_dev
+
 prune docs/_build
 prune build
