+++ astropy/_dev/__init__.py
@@ -1,0 +1,6 @@
+"""
+This package contains utilities that are only used when developing astropy
+in a copy of the source repository.
+
+These files are not installed, and should not be assumed to exist at runtime.
+"""
