+++ setup.cfg
@@ -33,6 +33,9 @@ install_requires =
     pyerfa
 python_requires = >=3.6
 
+[options.packages.find]
+exclude = astropy._dev
+
 [options.entry_points]
 console_scripts =
     fits2bitmap = astropy.visualization.scripts.fits2bitmap:main
