+++ setup.py
@@ -59,41 +59,6 @@ if 'build_docs' in sys.argv or 'build_sphinx' in sys.argv:
     print(DOCS_HELP)
     sys.exit(1)
 
-VERSION_TEMPLATE = """
-# Note that we need to fall back to the hard-coded version if either
-# setuptools_scm can't be imported or setuptools_scm can't determine the
-# version, so we catch the generic 'Exception'.
-try:
-    from setuptools_scm import get_version
-    version = get_version(root='..', relative_to=__file__)
-except Exception:
-    version = '{version}'
-else:
-    del get_version  # clean up namespace
-
-
-# We use LooseVersion to define major, minor, micro, but ignore any suffixes.
-def split_version(version):
-    pieces = [0, 0, 0]
-
-    try:
-        from distutils.version import LooseVersion
-
-        for j, piece in enumerate(LooseVersion(version).version[:3]):
-            pieces[j] = int(piece)
-
-    except Exception:
-        pass
-
-    return pieces
-
-
-major, minor, bugfix = split_version(version)
-
-del split_version  # clean up namespace.
-
-release = 'dev' not in version
-""".lstrip()
 
 # Only import these if the above checks are okay
 # to avoid masking the real problem with import error.
@@ -101,6 +66,5 @@ import os  # noqa
 from setuptools import setup  # noqa
 from extension_helpers import get_extensions  # noqa
 
-setup(use_scm_version={'write_to': os.path.join('astropy', 'version.py'),
-                       'write_to_template': VERSION_TEMPLATE},
+setup(use_scm_version={'write_to': os.path.join('astropy', '_version.py')},
       ext_modules=get_extensions())
